@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.29.1
+// 	protoc-gen-go v1.31.0
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/volume.proto
 
