@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -149,13 +149,12 @@ func (_m *ClusterServiceClient) Usage(ctx context.Context, in *v1.ClusterUsageRe
 	return r0, r1
 }
 
-type mockConstructorTestingTNewClusterServiceClient interface {
+// NewClusterServiceClient creates a new instance of ClusterServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewClusterServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewClusterServiceClient creates a new instance of ClusterServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewClusterServiceClient(t mockConstructorTestingTNewClusterServiceClient) *ClusterServiceClient {
+}) *ClusterServiceClient {
 	mock := &ClusterServiceClient{}
 	mock.Mock.Test(t)
 
