@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -118,13 +118,12 @@ func (_m *ClusterServiceServer) Usage(_a0 context.Context, _a1 *v1.ClusterUsageR
 	return r0, r1
 }
 
-type mockConstructorTestingTNewClusterServiceServer interface {
+// NewClusterServiceServer creates a new instance of ClusterServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewClusterServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewClusterServiceServer creates a new instance of ClusterServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewClusterServiceServer(t mockConstructorTestingTNewClusterServiceServer) *ClusterServiceServer {
+}) *ClusterServiceServer {
 	mock := &ClusterServiceServer{}
 	mock.Mock.Test(t)
 
