@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -149,13 +149,12 @@ func (_m *IPServiceClient) Usage(ctx context.Context, in *v1.IPUsageRequest, opt
 	return r0, r1
 }
 
-type mockConstructorTestingTNewIPServiceClient interface {
+// NewIPServiceClient creates a new instance of IPServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewIPServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewIPServiceClient creates a new instance of IPServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewIPServiceClient(t mockConstructorTestingTNewIPServiceClient) *IPServiceClient {
+}) *IPServiceClient {
 	mock := &IPServiceClient{}
 	mock.Mock.Test(t)
 
