@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -83,13 +83,12 @@ func (_m *InfoServiceClient) Tenants(ctx context.Context, in *v1.TenantInfoReque
 	return r0, r1
 }
 
-type mockConstructorTestingTNewInfoServiceClient interface {
+// NewInfoServiceClient creates a new instance of InfoServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewInfoServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewInfoServiceClient creates a new instance of InfoServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewInfoServiceClient(t mockConstructorTestingTNewInfoServiceClient) *InfoServiceClient {
+}) *InfoServiceClient {
 	mock := &InfoServiceClient{}
 	mock.Mock.Test(t)
 
