@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -66,13 +66,12 @@ func (_m *InfoServiceServer) Tenants(_a0 context.Context, _a1 *v1.TenantInfoRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewInfoServiceServer interface {
+// NewInfoServiceServer creates a new instance of InfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewInfoServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewInfoServiceServer creates a new instance of InfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewInfoServiceServer(t mockConstructorTestingTNewInfoServiceServer) *InfoServiceServer {
+}) *InfoServiceServer {
 	mock := &InfoServiceServer{}
 	mock.Mock.Test(t)
 
