@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -83,13 +83,12 @@ func (_m *NetworkTrafficServiceClient) Usage(ctx context.Context, in *v1.Network
 	return r0, r1
 }
 
-type mockConstructorTestingTNewNetworkTrafficServiceClient interface {
+// NewNetworkTrafficServiceClient creates a new instance of NetworkTrafficServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewNetworkTrafficServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewNetworkTrafficServiceClient creates a new instance of NetworkTrafficServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewNetworkTrafficServiceClient(t mockConstructorTestingTNewNetworkTrafficServiceClient) *NetworkTrafficServiceClient {
+}) *NetworkTrafficServiceClient {
 	mock := &NetworkTrafficServiceClient{}
 	mock.Mock.Test(t)
 
