@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -66,13 +66,12 @@ func (_m *NetworkTrafficServiceServer) Usage(_a0 context.Context, _a1 *v1.Networ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewNetworkTrafficServiceServer interface {
+// NewNetworkTrafficServiceServer creates a new instance of NetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewNetworkTrafficServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewNetworkTrafficServiceServer creates a new instance of NetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewNetworkTrafficServiceServer(t mockConstructorTestingTNewNetworkTrafficServiceServer) *NetworkTrafficServiceServer {
+}) *NetworkTrafficServiceServer {
 	mock := &NetworkTrafficServiceServer{}
 	mock.Mock.Test(t)
 
