@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -149,13 +149,12 @@ func (_m *PodServiceClient) Usage(ctx context.Context, in *v1.ContainerUsageRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPodServiceClient interface {
+// NewPodServiceClient creates a new instance of PodServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPodServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPodServiceClient creates a new instance of PodServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPodServiceClient(t mockConstructorTestingTNewPodServiceClient) *PodServiceClient {
+}) *PodServiceClient {
 	mock := &PodServiceClient{}
 	mock.Mock.Test(t)
 
