@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -118,13 +118,12 @@ func (_m *PodServiceServer) Usage(_a0 context.Context, _a1 *v1.ContainerUsageReq
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPodServiceServer interface {
+// NewPodServiceServer creates a new instance of PodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPodServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPodServiceServer creates a new instance of PodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPodServiceServer(t mockConstructorTestingTNewPodServiceServer) *PodServiceServer {
+}) *PodServiceServer {
 	mock := &PodServiceServer{}
 	mock.Mock.Test(t)
 
