@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -149,13 +149,12 @@ func (_m *PostgresServiceClient) Usage(ctx context.Context, in *v1.PostgresUsage
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPostgresServiceClient interface {
+// NewPostgresServiceClient creates a new instance of PostgresServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPostgresServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPostgresServiceClient creates a new instance of PostgresServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPostgresServiceClient(t mockConstructorTestingTNewPostgresServiceClient) *PostgresServiceClient {
+}) *PostgresServiceClient {
 	mock := &PostgresServiceClient{}
 	mock.Mock.Test(t)
 
