@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -118,13 +118,12 @@ func (_m *PostgresServiceServer) Usage(_a0 context.Context, _a1 *v1.PostgresUsag
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPostgresServiceServer interface {
+// NewPostgresServiceServer creates a new instance of PostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPostgresServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPostgresServiceServer creates a new instance of PostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPostgresServiceServer(t mockConstructorTestingTNewPostgresServiceServer) *PostgresServiceServer {
+}) *PostgresServiceServer {
 	mock := &PostgresServiceServer{}
 	mock.Mock.Test(t)
 
