@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -83,13 +83,12 @@ func (_m *S3ServiceClient) Usage(ctx context.Context, in *v1.S3UsageRequest, opt
 	return r0, r1
 }
 
-type mockConstructorTestingTNewS3ServiceClient interface {
+// NewS3ServiceClient creates a new instance of S3ServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewS3ServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewS3ServiceClient creates a new instance of S3ServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewS3ServiceClient(t mockConstructorTestingTNewS3ServiceClient) *S3ServiceClient {
+}) *S3ServiceClient {
 	mock := &S3ServiceClient{}
 	mock.Mock.Test(t)
 
