@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -66,13 +66,12 @@ func (_m *S3ServiceServer) Usage(_a0 context.Context, _a1 *v1.S3UsageRequest) (*
 	return r0, r1
 }
 
-type mockConstructorTestingTNewS3ServiceServer interface {
+// NewS3ServiceServer creates a new instance of S3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewS3ServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewS3ServiceServer creates a new instance of S3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewS3ServiceServer(t mockConstructorTestingTNewS3ServiceServer) *S3ServiceServer {
+}) *S3ServiceServer {
 	mock := &S3ServiceServer{}
 	mock.Mock.Test(t)
 
