@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeClusterServiceServer) mustEmbedUnimplementedClusterServiceServer
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeClusterServiceServer interface {
+// NewUnsafeClusterServiceServer creates a new instance of UnsafeClusterServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeClusterServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeClusterServiceServer creates a new instance of UnsafeClusterServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeClusterServiceServer(t mockConstructorTestingTNewUnsafeClusterServiceServer) *UnsafeClusterServiceServer {
+}) *UnsafeClusterServiceServer {
 	mock := &UnsafeClusterServiceServer{}
 	mock.Mock.Test(t)
 
