@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeIPServiceServer) mustEmbedUnimplementedIPServiceServer() {
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeIPServiceServer interface {
+// NewUnsafeIPServiceServer creates a new instance of UnsafeIPServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeIPServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeIPServiceServer creates a new instance of UnsafeIPServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeIPServiceServer(t mockConstructorTestingTNewUnsafeIPServiceServer) *UnsafeIPServiceServer {
+}) *UnsafeIPServiceServer {
 	mock := &UnsafeIPServiceServer{}
 	mock.Mock.Test(t)
 
