@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeInfoServiceServer) mustEmbedUnimplementedInfoServiceServer() {
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeInfoServiceServer interface {
+// NewUnsafeInfoServiceServer creates a new instance of UnsafeInfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeInfoServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeInfoServiceServer creates a new instance of UnsafeInfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeInfoServiceServer(t mockConstructorTestingTNewUnsafeInfoServiceServer) *UnsafeInfoServiceServer {
+}) *UnsafeInfoServiceServer {
 	mock := &UnsafeInfoServiceServer{}
 	mock.Mock.Test(t)
 
