@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeNetworkTrafficServiceServer) mustEmbedUnimplementedNetworkTraffi
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeNetworkTrafficServiceServer interface {
+// NewUnsafeNetworkTrafficServiceServer creates a new instance of UnsafeNetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeNetworkTrafficServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeNetworkTrafficServiceServer creates a new instance of UnsafeNetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeNetworkTrafficServiceServer(t mockConstructorTestingTNewUnsafeNetworkTrafficServiceServer) *UnsafeNetworkTrafficServiceServer {
+}) *UnsafeNetworkTrafficServiceServer {
 	mock := &UnsafeNetworkTrafficServiceServer{}
 	mock.Mock.Test(t)
 
