@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafePostgresServiceServer) mustEmbedUnimplementedPostgresServiceServ
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafePostgresServiceServer interface {
+// NewUnsafePostgresServiceServer creates a new instance of UnsafePostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafePostgresServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafePostgresServiceServer creates a new instance of UnsafePostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafePostgresServiceServer(t mockConstructorTestingTNewUnsafePostgresServiceServer) *UnsafePostgresServiceServer {
+}) *UnsafePostgresServiceServer {
 	mock := &UnsafePostgresServiceServer{}
 	mock.Mock.Test(t)
 
