@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeVolumeServiceServer) mustEmbedUnimplementedVolumeServiceServer()
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeVolumeServiceServer interface {
+// NewUnsafeVolumeServiceServer creates a new instance of UnsafeVolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeVolumeServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeVolumeServiceServer creates a new instance of UnsafeVolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeVolumeServiceServer(t mockConstructorTestingTNewUnsafeVolumeServiceServer) *UnsafeVolumeServiceServer {
+}) *UnsafeVolumeServiceServer {
 	mock := &UnsafeVolumeServiceServer{}
 	mock.Mock.Test(t)
 
