@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -149,13 +149,12 @@ func (_m *VolumeServiceClient) Usage(ctx context.Context, in *v1.VolumeUsageRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewVolumeServiceClient interface {
+// NewVolumeServiceClient creates a new instance of VolumeServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewVolumeServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewVolumeServiceClient creates a new instance of VolumeServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewVolumeServiceClient(t mockConstructorTestingTNewVolumeServiceClient) *VolumeServiceClient {
+}) *VolumeServiceClient {
 	mock := &VolumeServiceClient{}
 	mock.Mock.Test(t)
 
