@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.35.2. DO NOT EDIT.
 
 package mocks
 
@@ -118,13 +118,12 @@ func (_m *VolumeServiceServer) Usage(_a0 context.Context, _a1 *v1.VolumeUsageReq
 	return r0, r1
 }
 
-type mockConstructorTestingTNewVolumeServiceServer interface {
+// NewVolumeServiceServer creates a new instance of VolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewVolumeServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewVolumeServiceServer creates a new instance of VolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewVolumeServiceServer(t mockConstructorTestingTNewVolumeServiceServer) *VolumeServiceServer {
+}) *VolumeServiceServer {
 	mock := &VolumeServiceServer{}
 	mock.Mock.Test(t)
 
